Backend/requirements.txt
Flask
Pillow
numpy
watson-machine-learning-client
scikit-learn